Applied fix for bug #1794
Monotone-Parent: 3726e7db829b5ee868cdc21f61798d973f1482d4
Monotone-Revision: 87432dfeaa5feaee87a4ed6d05a7bb339c095278

Monotone-Author: ludovic@Sophos.ca
Monotone-Date: 2012-05-09T08:54:14
Monotone-Branch: ca.inverse.sope

+++ sope-xml/SaxObjC/SaxAttributes.m
@@ -185,7 +185,7 @@
   return [self valueAtIndex:i];
 }
 - (NSString *)valueForName:(NSString *)_localName uri:(NSString *)_uri {
-  unsigned int i;
+  NSUInteger i;
   
   if ((i = [self indexOfName:_localName uri:_uri]) == NSNotFound)
     return nil;
